requirements.txt
aws-cdk-lib==2.202.0
constructs>=10.0.0,<11.0.0
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.3.7
boto3==1.34.0
langgraph==0.0.20
langchain==0.1.0
langchain-openai==0.0.2
aiohttp==3.9.1
